Consistently use double for ranges (Issue #79)
Instead of mixing float/double, which leads to comparison failure

+++ lib/digitizer_block_impl.h
@@ -128,7 +128,7 @@ namespace gr {
           coupling(AC_1M)
       {}
 
-      float range;
+      double range;
       float offset;
       bool enabled;
       coupling_t coupling;
